issue13: add content service, api and format utils
