+++ src/components/ProtectedRoute/index.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Redirect, Route, RouteProps } from "react-router-dom";
 import config from "../../../docusaurus.config";
+import { isJsonFormat } from "../../utils/formatUtils";
 
 interface ProtectedRouteProps extends RouteProps {
   children: React.ReactNode;
@@ -10,19 +11,6 @@ const ProtectedRoute: React.FC<ProtectedRouteProps> = ({
   children,
   ...rest
 }) => {
-  const isJsonFormat = (text: string) => {
-    try {
-      const tokens = JSON.parse(text);
-
-      return (
-        tokens.hasOwnProperty("access") && tokens.hasOwnProperty("refresh")
-      );
-    } catch (error) {
-      localStorage.removeItem("jwt");
-      return false;
-    }
-  };
-
   const tokens: string | null = localStorage.getItem("jwt");
   const isAuthenticated: boolean = !!tokens && isJsonFormat(tokens);
