+++ src/utils/apiUtil.ts
@@ -1,0 +1,3 @@
+export const getBaseUrl = () => {
+  return "http://127.0.0.1:8000";
+};
